+++ pages/api/quotes/home.js
@@ -20,7 +20,6 @@ const QuotesHomeScraper = async (req, res) => {
       )
         .map((i, el) => {
           const $el = $(el);
-
           const imgURL = $el
             .find(" div.quoteDetails > a.leftAlignedImage")
             .attr("href");
@@ -63,6 +62,21 @@ const QuotesHomeScraper = async (req, res) => {
         })
         .toArray();
 
+      const popularTags = $(
+        "div.bigBoxContent.containerWithHeaderContent > ul.listTagsTwoColumn > li.greyText"
+      )
+        .map((i, el) => {
+          const $el = $(el);
+          const url = $el.find("a.gr-hyperlink").attr("href");
+          const name = $el.find("a.gr-hyperlink").text();
+          const id = i + 1;
+          return {
+            id: id,
+            url: url,
+            name: name,
+          };
+        })
+        .toArray();
       const lastScraped = new Date().toISOString();
       res.statusCode = 200;
       return res.json({
@@ -71,6 +85,7 @@ const QuotesHomeScraper = async (req, res) => {
         scrapeURL: scrapeURL,
         name: name,
         quotes: quotes,
+        popularTags: popularTags,
         lastScraped: lastScraped,
       });
     } catch (error) {