+++ pages/quotes/[...slug].js
@@ -10,6 +10,7 @@ const Slug = () => {
   const router = useRouter();
   const { slug } = router.query;
   const [scrapedData, setScrapedData] = useState({});
+  const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState(false);
 
   useEffect(() => {
@@ -30,8 +31,30 @@ const Slug = () => {
         setError(true);
       }
     };
+
+    const fetchTagData = async () => {
+      setIsLoading(true);
+      const res = await fetch(`/api/quotes/slug`, {
+        method: "POST",
+        headers: {
+          "content-type": "application/json",
+        },
+        body: JSON.stringify({
+          queryURL: `https://www.goodreads.com/quotes${`/tag?id=${router.query.id}`}`,
+        }),
+      });
+      if (res.ok) {
+        const data = await res.json();
+        setScrapedData(data);
+        setIsLoading(false);
+      } else {
+        setError(true);
+      }
+    };
     if (slug) {
-      fetchData();
+      {
+        !router.query.id ? fetchData() : fetchTagData();
+      }
     }
   }, [slug]);
 
@@ -40,6 +63,7 @@ const Slug = () => {
       <div className="bg-gradient-to-tr from-rose-50 to-rose-200 dark:bg-gradientedge text-gray-900 dark:text-gray-100 min-h-screen">
         <Header />
         {/* Show loader or error based on the scraper data response */}
+        {isLoading && <Loader other={true} />}
         {error && (
           <ErrorMessage
             status="500"
@@ -55,7 +79,7 @@ const Slug = () => {
                 url={`https://www.goodreads.com/quotes/${slug}`}
               />
             )}
-            {scrapedData.quotes && scrapedData.quotes.length === 0 && (
+            {scrapedData.quotes && scrapedData.quotes === "" && (
               <ErrorMessage
                 status="ScraperError"
                 url={`https://www.goodreads.com/quotes/${slug}`}