+++ pages/search/[...slug].js
@@ -5,8 +5,9 @@ import Header from "../../components/global-components/Header";
 import Footer from "../../components/global-components/Footer";
 import Loader from "../../components/global-components/Loader";
 import ErrorMessage from "../../components/global-components/ErrorMessage";
-import SearchResultData from "../../components/searchpage-components/SearchResultData";
+import BookResultData from "../../components/searchpage-components/BookResultData";
 import SearchBox from "../../components/searchpage-components/SearchBox";
+import QuotesResultData from "../../components/searchpage-components/QuotesResultData";
 
 const Slug = () => {
   const router = useRouter();
@@ -18,31 +19,25 @@ const Slug = () => {
   useEffect(() => {
     const scraperPath = searchType();
 
-    // TODO: Build new api search routes based on query type
+    // TODO: Add people, lists and groups search once those routes have been added
     function searchType() {
       if (query.type === "books") {
-        const scraperPath = "/api/search-scraper";
-        // const scraperPath = "/api/search/books";
+        const scraperPath = "/api/search/books";
         return scraperPath;
       } else if (query.type === "people") {
-        const scraperPath = "/api/search-scraper";
-        // const scraperPath = "/api/search/people"
+        const scraperPath = "/api/search/people";
         return scraperPath;
       } else if (query.type === "quotes") {
-        const scraperPath = "/api/search-scraper";
-        // const scraperPath = "/api/search/quotes"
+        const scraperPath = "/api/search/quotes";
         return scraperPath;
       } else if (query.type === "lists") {
-        const scraperPath = "/api/search-scraper";
-        // const scraperPath = "/api/search/lists"
+        const scraperPath = "/api/search/lists";
         return scraperPath;
       } else if (query.type === "groups") {
-        const scraperPath = "/api/search-scraper";
-        // const scraperPath = "/api/search/groups"
+        const scraperPath = "/api/search/groups";
         return scraperPath;
       } else {
-        const scraperPath = "/api/search-scraper";
-        // const scraperPath = "/api/search/books"
+        const scraperPath = "/api/search/books";
         return scraperPath;
       }
     }
@@ -107,6 +102,7 @@ const Slug = () => {
                 </div>
               </div>
             )}
+
             {scrapedData && (
               <>
                 {scrapedData.numberOfResults != "" && (
@@ -126,16 +122,24 @@ const Slug = () => {
 
                     <div className="flex justify-center items-center align-middle">
                       <div className="flex flex-col xl:flex-row justify-center items-center text-center w-full mt-8">
-                        <SearchBox />
+                        <SearchBox searchType={scrapedData.searchType} />
                       </div>
                     </div>
                   </div>
                 )}
-                <SearchResultData
-                  query={scrapedData.scrapeURL}
-                  result={scrapedData.result}
-                  numberOfResults={scrapedData.numberOfResults}
-                />
+                {scrapedData.searchType === "books" && (
+                  <BookResultData
+                    query={scrapedData.scrapeURL}
+                    result={scrapedData.result}
+                    numberOfResults={scrapedData.numberOfResults}
+                  />
+                )}
+                {/*  {scrapedData.searchType === "people" && "people"} */}
+                {scrapedData.searchType === "quotes" && (
+                  <QuotesResultData scrapedData={scrapedData} />
+                )}
+                {/*    {scrapedData.searchType === "lists" && "lists"} */}
+                {/*   {scrapedData.searchType === "groups" && "groups"} */}
                 {scrapedData.numberOfResults === "No results." && (
                   <div
                     id="noResults"